feat: Update CreatePatientTable component to include Pacemaker field
The CreatePatientTable component has been updated to include a Pacemaker field. This field allows the user to indicate whether the patient has a pacemaker or not. The field is implemented using a Select component from the Material-UI library, with options for "Yes" and "No". The selected value is stored in the formData state and can be accessed through the isPacemaker property. Validation rules have been added to ensure that the field is required. This change improves the functionality of the CreatePatientTable component by providing a way to capture the pacemaker status of a patient.