feat: Add Weeks Attended field to CreateSessionTable component
The CreateSessionTable component has been updated to include a Weeks Attended field. This field allows the user to indicate the number of weeks attended for a session. The field is implemented using a TextField component from the Material-UI library, with validation rules to ensure it is required. This change improves the functionality of the CreateSessionTable component by providing a way to capture the number of weeks attended for a session.